refactor(features/pairing): adjust component name
