+++ app/features/Pairing/Cards/PairingCard/CenterCard/constants.ts
