+++ app/features/Pairing/Cards/PairingCard/FrontCard/constants.ts
