+++ app/features/Pairing/Cards/index.tsx
@@ -1,6 +1,6 @@
-import CurrentCard from './CurrentCard';
+import PairingCard from './PairingCard';
 import RecommendCards from './RecommendCards';
-import { usePairContext } from '~/features/Pairing/context';
+import { usePairContext } from '~/features/pairing/context';
 
 export default function PairCards() {
   const { showPanel } = usePairContext();
@@ -13,7 +13,7 @@ export default function PairCards() {
       flex="col"
       {...(showPanel && { display: 'none lg:flex' })}
     >
-      <CurrentCard />
+      <PairingCard />
       <RecommendCards />
     </section>
   );
