+++ app/features/Pairing/Cards/utils.ts
