+++ app/features/Pairing/ControlPanel/Filter/Item/type.d.ts
@@ -7,12 +7,12 @@ import type {
   AgeOptions,
   Color,
   ColorOptions
-} from '~/features/Pairing/Panel/Filter/type';
+} from '~/features/pairing/ControlPanel/Filter/type';
 
 import {
   GENERAL_FILTER_OPTIONS,
   DYNAMIC_FILTER_OPTIONS
-} from '~/features/Pairing/Panel/Filter/constants';
+} from '~/features/pairing/ControlPanel/Filter/constants';
 
 export type Category =
   | typeof GENERAL_FILTER_OPTIONS[number]['CATEGORY']
@@ -23,13 +23,3 @@ export type Label =
   | typeof DYNAMIC_FILTER_OPTIONS[number]['LABEL'];
 
 export type Options = FamilyOptions | GenderOptions | AgeOptions | ColorOptions;
-
-type Filter = {
-  family: Family;
-  gender: Gender;
-  age: Age;
-  color: Color;
-};
-
-type FilterType = keyof Filter;
-type Payload = ValueOf<Filter>;
