+++ app/features/Pairing/ControlPanel/Filter/index.tsx
@@ -1,6 +1,6 @@
 import { useReducer } from 'react';
 import type { Family, Gender, Age, Color } from './type';
-import FilterItem from '~/features/Pairing/Panel/FilterItem';
+import FilterItem from '~/features/pairing/ControlPanel/Filter/Item';
 import {
   DEFAULT_OPTION,
   GENERAL_FILTER_OPTIONS,
