+++ app/features/Pairing/ControlPanel/Setting/index.tsx
