+++ app/features/Pairing/index.tsx
@@ -1,7 +1,7 @@
-import Panel from './Panel';
-import PairCards from './PairCards';
-import Portal from './Portal';
 import Loading from '~/components/common/LoadingAnimation';
+import ControlPanel from './ControlPanel';
+import PairCards from './Cards';
+import Portal from './Portal';
 import { PairingContextProvider } from './context';
 
 export default function Pairing() {
@@ -14,7 +14,7 @@ export default function Pairing() {
         h="[calc(100vh-160px)] sm:[calc(100vh-120px)] lg:[calc(100vh-80px)]"
       >
         <Portal />
-        <Panel />
+        <ControlPanel />
         {isLoading ? <Loading /> : <PairCards />}
       </main>
     </PairingContextProvider>
