+++ app/routes/index.tsx
@@ -1,5 +1,5 @@
 import { useEffect, useState } from 'react';
-import Pairing from '~/features/Pairing';
+import Pairing from '~/features/pairing';
 import FullPageLoading from '~/components/common/FullPageLoading';
 
 const LOADING_SECONDS = 1000 * 1.5;
